requirements.txt
numpy~=1.22.3
matplotlib~=3.5.1
pillow~=9.0.1
scipy~=1.8.0
pytest~=7.1.1
setuptools==60.10.0
cython~=0.29.28